requirements.txt
discord.py
genshin
aiohttp
python-dotenv
Pillow>=9.5.0
imageio
imageio-ffmpeg
requests
psutil
